feat(comments): add ability to comment on a season
Currently this is only restricted to a logged in user.
Comments can only be updated by their owner but can be "deleted"
(ie hidden) by the season owner as well (currently this is just done
through a boolean flag / client hiding, but it may be switched to server-side later).

Conflicts:
	bower.json
	frontend/app.coffee
	frontend/scss/app.scss
	public/index.html